wip: restructure repo
