+++ src/earthkit/hydro/core/__init__.py
@@ -1,0 +1,2 @@
+from .accumulate import flow_downstream, flow_upstream
+from .flow import propagate
