+++ src/earthkit/hydro/data_structures/__init__.py
@@ -1,0 +1,1 @@
+from .network import RiverNetwork
