+++ src/earthkit/hydro/distance/__init__.py
@@ -1,0 +1,1 @@
+from .metrics import max, min, to_sink, to_source
