+++ src/earthkit/hydro/downstream/operations.py
@@ -1,0 +1,165 @@
+import numpy as np
+
+from earthkit.hydro.core.online import calculate_online_metric
+from earthkit.hydro.utils import mask_and_unmask
+
+
+def calculate_downstream_metric(
+    river_network,
+    field,
+    metric,
+    node_weights=None,
+    edge_weights=None,
+    mv=np.nan,
+    accept_missing=False,
+):
+    return calculate_online_metric(
+        river_network,
+        field,
+        metric,
+        node_weights,
+        edge_weights,
+        mv,
+        accept_missing,
+        flow_direction="up",
+    )
+
+
+@mask_and_unmask
+def var(
+    river_network,
+    field,
+    node_weights=None,
+    edge_weights=None,
+    mv=np.nan,
+    accept_missing=False,
+):
+    return calculate_downstream_metric(
+        river_network,
+        field,
+        "var",
+        node_weights,
+        edge_weights,
+        mv,
+        accept_missing,
+    )
+
+
+@mask_and_unmask
+def std(
+    river_network,
+    field,
+    node_weights=None,
+    edge_weights=None,
+    mv=np.nan,
+    accept_missing=False,
+):
+    return calculate_downstream_metric(
+        river_network,
+        field,
+        "std",
+        node_weights,
+        edge_weights,
+        mv,
+        accept_missing,
+    )
+
+
+@mask_and_unmask
+def mean(
+    river_network,
+    field,
+    node_weights=None,
+    edge_weights=None,
+    mv=np.nan,
+    accept_missing=False,
+):
+    return calculate_downstream_metric(
+        river_network,
+        field,
+        "mean",
+        node_weights,
+        edge_weights,
+        mv,
+        accept_missing,
+    )
+
+
+@mask_and_unmask
+def min(
+    river_network,
+    field,
+    node_weights=None,
+    edge_weights=None,
+    mv=np.nan,
+    accept_missing=False,
+):
+    return calculate_downstream_metric(
+        river_network,
+        field,
+        "min",
+        node_weights,
+        edge_weights,
+        mv,
+        accept_missing,
+    )
+
+
+@mask_and_unmask
+def max(
+    river_network,
+    field,
+    node_weights=None,
+    edge_weights=None,
+    mv=np.nan,
+    accept_missing=False,
+):
+    return calculate_downstream_metric(
+        river_network,
+        field,
+        "max",
+        node_weights,
+        edge_weights,
+        mv,
+        accept_missing,
+    )
+
+
+@mask_and_unmask
+def sum(
+    river_network,
+    field,
+    node_weights=None,
+    edge_weights=None,
+    mv=np.nan,
+    accept_missing=False,
+):
+    return calculate_downstream_metric(
+        river_network,
+        field,
+        "sum",
+        node_weights,
+        edge_weights,
+        mv,
+        accept_missing,
+    )
+
+
+@mask_and_unmask
+def prod(
+    river_network,
+    field,
+    node_weights=None,
+    edge_weights=None,
+    mv=np.nan,
+    accept_missing=False,
+):
+    return calculate_downstream_metric(
+        river_network,
+        field,
+        "prod",
+        node_weights,
+        edge_weights,
+        mv,
+        accept_missing,
+    )
