+++ src/earthkit/hydro/move/__init__.py
@@ -1,0 +1,1 @@
+from .metrics import downstream, upstream
