+++ src/earthkit/hydro/move/metrics.py
@@ -1,19 +1,13 @@
-# (C) Copyright 2025- ECMWF.
-#
-# This software is licensed under the terms of the Apache Licence Version 2.0
-# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
-# In applying this licence, ECMWF does not waive the privileges and immunities
-# granted to it by virtue of its status as an intergovernmental organisation
-# nor does it submit to any jurisdiction.
-
 import numpy as np
 
-from .utils import mask_and_unmask, missing_to_nan, nan_to_missing
+from earthkit.hydro.data_structures import RiverNetwork
+from earthkit.hydro.utils.decorators import mask_and_unmask
+from earthkit.hydro.utils.missing import missing_to_nan, nan_to_missing
 
 
 @mask_and_unmask
-def move_downstream(
-    river_network,
+def downstream(
+    river_network: RiverNetwork,
     field,
     node_weights=None,
     edge_weights=None,
@@ -61,36 +55,25 @@ def move_downstream(
         edge_weights, _ = missing_to_nan(edge_weights.copy(), mv, accept_missing)
 
     ups = np.zeros(field.shape, dtype=field_dtype)
-    if not river_network.has_bifurcations:
-        mask = (
-            river_network.downstream_nodes != river_network.n_nodes
-        )  # remove sinks since they have no downstream
-        nodes_to_update = river_network.downstream_nodes[mask]
-        values_to_add = (
-            field[..., mask]
-            if node_weights is None
-            else field[..., mask] * node_weights[..., mask]
-        )
-    else:
-        nodes_to_update = river_network.down_ids
-        values_to_add = (
-            field[..., river_network.up_ids]
-            if node_weights is None
-            else field[..., river_network.up_ids]
-            * node_weights[..., river_network.up_ids]
-        )
+
+    did, uid, eid = river_network._storage.sorted_data
+    values_to_add = (
+        field[..., uid]
+        if node_weights is None
+        else field[..., uid] * node_weights[..., uid]
+    )
 
     if edge_weights is not None:
         values_to_add *= edge_weights
 
-    ufunc.at(ups, (*[slice(None)] * (ups.ndim - 1), nodes_to_update), values_to_add)
+    ufunc.at(ups, (*[slice(None)] * (ups.ndim - 1), did), values_to_add)
     ups = nan_to_missing(ups, field_dtype, mv)
     return ups
 
 
 @mask_and_unmask
-def move_upstream(
-    river_network,
+def upstream(
+    river_network: RiverNetwork,
     field,
     node_weights=None,
     edge_weights=None,
@@ -137,25 +120,16 @@ def move_upstream(
         edge_weights, _ = missing_to_nan(edge_weights.copy(), mv, accept_missing)
 
     down = np.zeros(field.shape, dtype=field_dtype)
-    if not river_network.has_bifurcations:
-        mask = river_network.downstream_nodes != river_network.n_nodes  # remove sinks
-        update_vals = field[..., river_network.downstream_nodes[mask]]
-        if node_weights is not None:
-            update_vals *= node_weights[..., river_network.downstream_nodes[mask]]
-        if edge_weights is not None:
-            update_vals *= edge_weights
-        down[..., mask] = update_vals
-    else:
-        nodes_to_update = river_network.up_ids
-        update_vals = (
-            field[..., river_network.down_ids]
-            if node_weights is None
-            else field[..., river_network.down_ids]
-            * node_weights[..., river_network.down_ids]
-        )
-        if edge_weights is not None:
-            update_vals *= edge_weights
-        ufunc.at(down, (*[slice(None)] * (down.ndim - 1), nodes_to_update), update_vals)
+
+    did, uid, eid = river_network._storage.sorted_data
+    update_vals = (
+        field[..., did]
+        if node_weights is None
+        else field[..., did] * node_weights[..., did]
+    )
+    if edge_weights is not None:
+        update_vals *= edge_weights
+    ufunc.at(down, (*[slice(None)] * (down.ndim - 1), uid), update_vals)
 
     down = nan_to_missing(down, field_dtype, mv)
