+++ src/earthkit/hydro/river_network/__init__.py
@@ -1,0 +1,1 @@
+from .river_network import available, create, load
