+++ src/earthkit/hydro/subcatchments/__init__.py
@@ -1,0 +1,2 @@
+from .find import find
+from .metrics import max, mean, min, prod, std, sum, var
