+++ src/earthkit/hydro/upstream/__init__.py
@@ -1,0 +1,1 @@
+from .operations import max, mean, min, prod, std, sum, var
