+++ src/earthkit/hydro/utils/__init__.py
@@ -1,0 +1,9 @@
+from .convert import points_to_1d_indices, points_to_numpy
+from .decorators import mask_2d, mask_and_unmask
+from .missing import (
+    are_missing_values_present,
+    check_missing,
+    is_missing,
+    missing_to_nan,
+    nan_to_missing,
+)
