+++ src/earthkit/hydro/zonal/__init__.py
@@ -1,0 +1,1 @@
+from .metrics import max, mean, min, prod, std, sum, var
